requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.25.0
requests>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
endee>=0.1.0
